fix bug with values for mars and jupiter.

+++ src/main.js
@@ -5,14 +5,12 @@ import 'bootstrap/dist/css/bootstrap.min.css';
 import './styles.css';
 
 $(document).ready(function() {
-
-
   $("#user-info").submit(function(event) {
     event.preventDefault();
     const userName = $('#user-name').val();
-    const userAge = parseInt($('#user-age').val());
-    const userDeathAge = parseInt($('#user-death-age').val());
-    let aging = parseInt($('#aging-selector').val());
+    const userAge = parseFloat($('#user-age').val());
+    const userDeathAge = parseFloat($('#user-death-age').val());
+    let aging = parseFloat($('#aging-selector').val());
     let planetName = $('#aging-selector option:selected').text();
     let user = new User(userName, userAge, userDeathAge, aging);
     if (userAge > user.death) {
@@ -27,7 +25,5 @@ $(document).ready(function() {
     $('.age-calc').show();
     $('.results').show();
     $('.user-entry').hide();
-
-
   });
 });
